refactor: extract CalendarHeader from DesktopLayout
Extract the sticky combined header (current-task banner, date headers
row, and all-day tasks section) into a CalendarHeader component. This
is desktop/tablet only and calls useDayPlannerCtx() directly.

DesktopLayout drops from ~2370 lines to ~1714 lines.
